requirements.txt
torch>=1.9.0
transformers>=4.20.0
requests>=2.28.0
beautifulsoup4>=4.11.0
newspaper3k>=0.2.8
google-search-results>=2.4.2
newsapi-python>=0.2.6
pytesseract>=0.3.10
opencv-python>=4.6.0
Pillow>=9.2.0
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0
spacy>=3.4.0
textblob>=0.17.1
python-dotenv>=0.19.0
fastapi>=0.78.0
uvicorn>=0.18.0
pydantic>=1.9.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
sentence-transformers>=2.2.0
